pyproject.toml
[project]
name = "adrf-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "adrf>=0.1.9",
    "daphne==4.2.1",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
]
